requirements.txt
alembic==1.4.2
anyjson==0.3.3
apache-airflow==1.10.10
apispec==1.3.3
argcomplete==1.12.0
argon2-cffi==20.1.0
arrow==0.15.8
attrs==19.3.0
Automat==0.8.0
autopep8==1.4.4
Babel==2.8.0
bcrypt==3.1.7
beautifulsoup4==4.8.1
boto3==1.9.66
botocore==1.12.189
brotlipy==0.7.0
cached-property==1.5.1
cattrs==1.0.0
certifi==2020.6.20
cffi==1.13.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colorlog==4.2.1
colour==0.1.5
configparser==5.0.0
constantly==15.1.0
croniter==0.3.30
cryptography==2.8
cssselect==1.1.0
defusedxml==0.6.0
dill==0.3.2
docutils==0.16
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==2.2.1
Flask-Babel==0.12.2
Flask-Caching==1.7.1
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.13
Flask-WTF==0.14.3
funcsigs==1.0.2
furl==2.1.0
future==0.18.2
graphviz==0.14.1
gunicorn==20.0.4
hyperlink==17.3.1
idna==2.8
importlib-metadata==1.7.0
incremental==17.5.0
infinity==1.4
intervals==0.8.1
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
json-merge-patch==0.2
jsonschema==3.0.2
lazy-object-proxy==1.5.1
lockfile==0.12.2
lxml==4.4.1
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
more-itertools==7.2.0
numpy==1.19.1
orderedmultidict==1.0.1
packaging==19.2
pandas==1.1.0
parsel==1.5.2
passlib==1.7.2
pendulum==1.4.4
phonenumbers==8.10.14
pluggy==0.12.0
prison==0.1.3
Protego==0.1.15
psutil==5.7.2
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.6.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==19.0.0
pyparsing==2.4.5
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.3.0
pytest-runner==5.2
python-daemon==2.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python3-openid==3.1.0
pytz==2020.1
pytzdata==2020.1
PyYAML==5.3.1
queuelib==1.5.0
requests==2.24.0
s3transfer==0.1.13
Scrapy==1.8.0
service-identity==18.1.0
setproctitle==1.1.10
six==1.13.0
soupsieve==1.9.5
SQLAlchemy==1.3.18
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.6
tabulate==0.8.7
tenacity==6.2.0
termcolor==1.1.0
text-unidecode==1.3
thrift==0.13.0
Twisted==19.10.0
typing-extensions==3.7.4.2
tzlocal==2.1
unicodecsv==0.14.1
urllib3==1.25.9
w3lib==1.21.0
wcwidth==0.1.7
Werkzeug==0.16.1
WTForms==2.3.2
zipp==0.6.0
zope.deprecation==4.4.0
zope.interface==4.7.1
